requirements.txt
cytoolz==0.9.0
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
eth-utils==0.7.4
Markdown==2.6.11
pysha3==1.0.2
pytz==2018.3
toolz==0.9.0
